fix slash command stuff
